wip
Starts extracting test helper methods, though we have a long way to go.
I think we should have some type of Suite struct, like we do for the
AsyncInserter, where each suite provides a function that can instantiate
the harness.

Something to help us say "when I write this to Postgres, I expect to see
this" in a sink agnostic way is going to be really important too.